refactor, inline edit save only when there are changes detected
