+++ BlazorAwesome/Models/AweUtil.cs
@@ -2,46 +2,12 @@
 using System.Collections.Generic;
 using System.Text.Json;
 using System.Linq;
-using System.Collections;
 using System;
 using System.Threading.Tasks;
 using Microsoft.JSInterop;
-using Microsoft.JSInterop.Implementation;
-using System.Text;
-using System.Reflection;
-using Microsoft.AspNetCore.Components;
 
 namespace Omu.BlazorAwesome.Models
 {
-    internal static class JSUtil
-    {
-        public static async Task<IJSObjectReference> InvokeAsync(IJSRuntime JS, string identifier, object obj)
-        {
-            try
-            {
-                return await JS.InvokeAsync<IJSObjectReference>(identifier, obj);
-            }
-            catch (ObjectDisposedException)
-            {
-            }
-
-            return null;
-        }
-
-        public static async Task InvokeVoidAsync(IJSRuntime JS, string identifier, object obj)
-        {
-            try
-            {
-                await JS.InvokeVoidAsync(identifier, obj);
-            }
-            catch (ObjectDisposedException)
-            {
-            }
-
-            return;
-        }
-    }
-
     /// <summary>
     /// Utils
     /// </summary>
