+++ UiServer/Models/Site/SiteMap.cs
@@ -12,7 +12,7 @@ namespace UiServer.Models.Site
         new("Popup", "Popup") { Keywords = "form open"},
         new("GridInlineEdit", "Grid Inline Editing") { Keywords = "crud"},
         new("GridNestCrud", "Grid Nest Crud"),
-        new("gridmasterdetailwithinlinedit", "Grid Master Detail With Inline Editing"),
+        new("GridMasterDetailWithInlinEdit", "Grid Master Detail With Inline Editing"),
         new("GridCrud", "Grid Popup Crud") { Keywords = "create edit delete"},
         new("GridFilterRow", "Grid Filter Row") { Keywords = "search"},
         new("GridGrouping", "Grid Grouping") { Keywords = "aggregates"},
