+++ UiServer/Pages/GridMasterDetailWithInlineEdit.razor
@@ -1,4 +1,4 @@
-﻿@page "/gridmasterdetailwithinlinedit"
+﻿@page "/GridMasterDetailWithInlinEdit"
 @using Omu.BlazorAwesome.Core
 @inject IDbContextFactory<MyContext> cxf
 <div class="expl">
